made the program into smaller chunks
